+++ src/Text/Seonbi/Html/Printer.hs
@@ -1,6 +1,8 @@
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 module Text.Seonbi.Html.Printer
     ( printHtml
+    , printText
     , printXhtml
     ) where
 
@@ -9,14 +11,19 @@ import Data.List
 
 import qualified Data.Text
 import Data.Text.Lazy
+import Data.Text.Lazy.Builder
+import HTMLEntities.Decoder
 
 import Text.Seonbi.Html.Entity
 import Text.Seonbi.Html.Tag
 
--- | Print the list of 'HtmlEntity' into a lazy 'Text'.
---
+-- $setup
 -- >>> :set -XOverloadedStrings
 -- >>> import Text.Seonbi.Html.Scanner
+-- >>> :set -interactive-print=Text.Show.Unicode.uprint
+
+-- | Print the list of 'HtmlEntity' into a lazy 'Text'.
+--
 -- >>> let Done "" tokens = scanHtml "<p>Hello,<br>\n<em>world</em>!</p>"
 -- >>> printHtml tokens
 -- "<p>Hello,<br>\n<em>world</em>!</p>"
@@ -26,8 +33,6 @@ printHtml = printHtml' False
 -- | Similar to 'printHtml' except it renders void (self-closing) tags as
 -- like @<br/>@ instead of @<br>@.
 --
--- >>> :set -XOverloadedStrings
--- >>> import Text.Seonbi.Html.Scanner
 -- >>> let Done "" tokens = scanHtml "<p>Hello,<br>\n<em>world</em>!</p>"
 -- >>> printXhtml tokens
 -- "<p>Hello,<br/>\n<em>world</em>!</p>"
@@ -72,3 +77,26 @@ printHtml' xhtml =
     renderAttrs attrs
       | isSpace (Data.Text.head attrs) = fromStrict attrs
       | otherwise = ' ' `cons` fromStrict attrs
+
+-- | Print only the text contents (including CDATA sections) without tags
+-- into a lazy 'Text'.
+--
+-- >>> let Done "" tokens = scanHtml "<p>Hello,<br>\n<em>world</em>!</p>"
+-- >>> printText tokens
+-- "Hello,\nworld!"
+--
+-- Entities are decoded:
+--
+-- >>> let Done "" tokens = scanHtml "<p><code>&lt;&gt;&quot;&amp;</code></p>"
+-- >>> printText tokens
+-- "<>\"&"
+printText :: [HtmlEntity] -> Text
+printText [] = Data.Text.Lazy.empty
+printText (x:xs) =
+    render x <> printText xs
+  where
+    render :: HtmlEntity -> Text
+    render = \ case
+        HtmlText { rawText = t } -> toLazyText $ htmlEncodedText t
+        HtmlCdata { text = t } -> fromStrict t
+        _ -> Data.Text.Lazy.empty